1.Use Flask BluePrint, separating site route and api route 2.Use dotenv to keep important data
